refactor: get > getSwapProposal
Also refactored the create endpoint files to keep consistency
